Improv: move TLV to msgs.ts and add comments
